Reorganize tests to clarify scopes
